backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai>=1.51.0
vectorbt==0.26.2
pandas==2.2.3
numpy==1.26.4
httpx==0.28.1
supabase==2.11.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
PyNaCl>=1.5.0
base58>=2.1.0
solana>=0.34.0
solders>=0.21.0
Pillow==11.1.0
python-dotenv==1.0.1
slowapi==0.1.9
openai>=1.0.0
anthropic>=0.40.0
chromadb==1.5.1
pytest
pytest-asyncio
